endpoints restified

+++ backend-kibe/src/main/java/org/ajc2020/spring1/controller/WorkstationController.java
@@ -1,5 +1,7 @@
 package org.ajc2020.spring1.controller;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.ajc2020.spring1.model.Workstation;
 import org.ajc2020.spring1.service.WorkstationService;
 import org.ajc2020.utility.communication.WorkStationCreationRequest;
@@ -37,6 +39,11 @@ public class WorkstationController {
         return station;
     }
 
+    @Operation(
+            description = "Returns all workstations",
+            tags = "Workstations",
+            security = {@SecurityRequirement(name = "user", scopes = "admin")}
+    )
     @GetMapping
     public List<WorkStationResource> returnWorkstations() {
         return workstationService.findAll().stream()
@@ -45,6 +52,11 @@ public class WorkstationController {
                 .collect(Collectors.toList());
     }
 
+    @Operation(
+            description = "Create a workstation",
+            tags = "Workstations",
+            security = {@SecurityRequirement(name = "user", scopes = "admin")}
+    )
     @PostMapping
     public ResponseEntity<WorkStationResource> createWorkstation(@Valid @RequestBody WorkStationCreationRequest request) {
         Workstation workstation = new Workstation(request);
@@ -54,6 +66,11 @@ public class WorkstationController {
         );
     }
 
+    @Operation(
+            description = "Returns a workstation",
+            tags = "Workstations",
+            security = {@SecurityRequirement(name = "user", scopes = "admin")}
+    )
     @GetMapping("/{id}")
     public ResponseEntity<WorkStationResource> returnWorkstation(@PathVariable String id, Locale locale) {
         ResourceBundle resourceBundle = ResourceBundle.getBundle("messages", locale);
@@ -64,16 +81,26 @@ public class WorkstationController {
         );
     }
 
+    @Operation(
+            description = "Removes a workstation",
+            tags = "Workstations",
+            security = {@SecurityRequirement(name = "user", scopes = "admin")}
+    )
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> removeWorkstation(@PathVariable String id, Locale locale) {
         ResourceBundle resourceBundle = ResourceBundle.getBundle("messages", locale);
-        workstationService.findById(id)
-                .orElseThrow(() -> new WorkstationNotFound(resourceBundle.getString("error.workstation.not.found")));
+        if(!workstationService.findById(id).isPresent())
+            throw new WorkstationNotFound(resourceBundle.getString("error.workstation.not.found"));
         workstationService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
 
-    @PatchMapping("/{id}/kick")
+    @Operation(
+            description = "Removes the current user from the workstation",
+            tags = "Workstations",
+            security = {@SecurityRequirement(name = "user", scopes = "admin")}
+    )
+    @DeleteMapping("/{id}/occupier")
     public ResponseEntity<WorkStationResource> kickWorker(@PathVariable String id, Locale locale) {
         ResourceBundle resourceBundle = ResourceBundle.getBundle("messages", locale);
         Workstation workStation = workstationService.findById(id)
@@ -83,8 +110,12 @@ public class WorkstationController {
         return ResponseEntity.ok(workStation.toResource());
     }
 
-    // TODO: fix naming and mapping
-    @PatchMapping("/{id}/enable")
+    @Operation(
+            description = "Allows the use of the workstation",
+            tags = "Workstations",
+            security = {@SecurityRequirement(name = "user", scopes = "admin")}
+    )
+    @PostMapping("/{id}/enabled")
     public ResponseEntity<WorkStationResource> enableWorkstation(@PathVariable String id, Locale locale) {
         ResourceBundle resourceBundle = ResourceBundle.getBundle("messages", locale);
         Workstation workStation = workstationService.findById(id)
@@ -97,8 +128,12 @@ public class WorkstationController {
         return ResponseEntity.ok(workStation.toResource());
     }
 
-    // FIXME: maybe this needs to be put in one request
-    @PatchMapping("/{id}/disable")
+    @Operation(
+            description = "Disallows the use of the workstation",
+            tags = "Workstations",
+            security = {@SecurityRequirement(name = "user", scopes = "admin")}
+    )
+    @DeleteMapping("/{id}/enabled")
     public ResponseEntity<WorkStationResource> disableWorkstation(@PathVariable String id, Locale locale) {
         ResourceBundle resourceBundle = ResourceBundle.getBundle("messages", locale);
         Workstation workStation = workstationService.findById(id)
